+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - **Added:** Rule `DOC012` to check for leading in first non-blank line of docstring.
 - **Added:** Rule `DOC106` to check parameter order mismatch with function signature.
+- **Added:** Rule `DOC107` to check missing parameters in docstring. (by default disabled)
 - **Added:** Add CLI parameter `--version`
 - **Added:** Allow read custom configuration from `configuration from `pyproject.toml` files.`.